fix: Add initial divider with no padding to apply corners conditionally
There is no CSS selector combinator that allows conditional styling of an element conditional upon "subsequent" sibling elements existing. This is a work-around whereby a dummy divider with no padding is added to the start of the button group such that the custom corner overrides only apply to button groups that contain a 1st and 3rd element that are dividers. Only options for conditional styling dependent on "preceding" sibling elements currently exist that are compatible across browsers (i.e. 'Adjacent sibling combinator' by use of `+` and 'General sibling combinator' by use of `~`, see below reference for details).

Another alternative solution that was considered by not deemed generic enough was to conditionally apply the custom cornering only to button groups that are contained within a classname that contains the text "--Editor", on the assumption that the Forget, Reset, Save button group would only be used within a nesting containing such a wildcard value (as is currently the case)
```
div [class*="--Editor"] {
  .ui--Button-Group {
    .ui.buttons {
      .button:nth-child(1) {
        border-top-right-radius: $small-corner;
        border-bottom-right-radius: $small-corner;
      }
      .ui.button.compact.mini.basic + .button:nth-child(3) {
        border-top-left-radius: $small-corner;
        border-bottom-left-radius: $small-corner;
      }
    }
  }
}
```

Reference: https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Selectors