fix: Fixes #297. Toggles Creator menu tab before initial mount when no addresses exist
* Fixes bug introduced by PR #292 that didn't cater for switching to the Creator menu tab on initial mount when no accounts exist.
The user shouldn't need to be notified and have to manually switch menu tabs when there aren't any addresses, the UI should do the work for the user and switch automatically depending on the current state.
It only catered for updates to the available addresses (i.e. forgetting an address when it was the only one remaining would toggle dispalying the Creator menu tab)
https://github.com/polkadot-js/apps/pull/292