fix: Move restoreAccount logic into ui-keyring. Fix bug. Remove addAccountPair. See detailed comments
IMPORTANT NOTE: A bug was identified, hence the reason for console.error being added to catch blocks in Backup.tsx and Restore.tsx for debugging purposes. To reproduce the error in the code prior to this commit:

* Forget all accounts
* Create an account with password
* Backup the account (works the first time you download it)
* Restore the account by selecting the JSON file that was just downloaded
* Backup the same account (for the second time) does not work. View Chrome Inspector to view error ExtError: Invalid Pkcs8 divider found in body that is caught in Backup.tsx due to calling pair.decodePkcs8 in backup.ts

Note that I hadn't encountered this same bug using the same workflow in my luke-74-download-upload-with-password branch, so I comparing the code closely to identify the issue.

It was found to be caused in PR #336 by the call to keyring.addAccountPair(pair, password); in Restore.tsx, specifically where in addAccountPair it calls saveAccount and obtains the json again const json = pair.toJson(password); and then calls state.keyring.addFromJson(json); with that json, instead of using the original json from onSave in Restore.tsx.