WIP: Fixes #417 but does not cater for #72.
* Fixes #417, but only caters for Chain State queries with 0 or 1 parameter,
for >=2 parameters to be catered for we need to address #72

* Incomplete is resolving TSLint errors
(i.e. `Argument of type 'RawParam & TypeDef' is not assignable to parameter of type 'MethodParams'.
Property 'params' is missing in type 'RawParam & TypeDef'.`) when trying to combine the method
parameter's value and type into a single object (since the method parameter's value and type are
passed into the `onAdd` separately contained in the `values` and `params` of the `this.state`)

* Incomplete is knowing whether to apply the `isValid` property further.

* Incomplete is checking the functionality is correct in the "Raw key" tab of http://localhost:3000/#/chainstate
(where a hex-encoded storage key is provided) since functionality has only been tested in the "Module" tab.