Fix translations loading for ever
Currently when going to #/settings/i18n we have the spinner loading forever.

It is caused by 404 on locales/pl/translation.json missing.

bug introduced by this PR:
https://github.com/polkadot-js/apps/pull/4724

To avoid this in future and instead of creating a locales/pl/translation.json (that can be done in another PR) ii catch any error on retrieveJson and return an empty {}