Use try-catch instead of check for null
The check for `null` at the start of the `Data` function meant that an `Option` that `isNone` would return `()`. This was because `value` for `Option.isNone == true` is `null`.

Instead, we can wrap all the type handlers in a try-catch, and let `createTypeUnsafe` in `formatData` deal with our invalid values (and throw if something bad happens).

Apologies for the large number of changed lines - the try-catch required an additional level of indentation.