Improving UX when making balance transfers
Although the front end can stop the user from sending a transaction by checking if the balance transfer will kill the account, it sends it to the network! 

Here is an instance that shows how detrimental it is for UX - https://polkadot.subscan.io/account/15vzAirFFFWcEgF74jxLoCTQxh11caDYgBrdQpUAnSEaDkrz

With some additional code logic, this can be averted. Also, the corner case for the second issue mentioned here is not handled well - https://github.com/polkadot-js/apps/issues/4931

This PR will take some time to finish. Need to understand how things work, before any changes are pushed to the PR.