Update testingRelayRococo.ts
Hi,
the ParaID 3025 is now registered for helixstreet (historically Arctic). The rpc-endpoint is live; the parachain didn't connect yet in the current slot due to issues with the genesis block - will recreate WASM and Genesis Files. 

The Arctic entry is here removed and the helixstreet entry entered alphabetically based on the text field.