Update productionRelayPolkadot.ts
We faced a DNS issue with one of our onprem servers that confused the front end load balancer causing a failure with the statemint endpoint. We have been working to fix it. The fix is in place and endpoint has been operating stable for the day. We continue to monitor it with high priority.