Update productionRelayPolkadot.ts
Addressing https://github.com/polkadot-js/apps/issues/9480

We have switched to using the relay chain as external RPC rather than running a relay chain RPC along with asset chains - to take advantage of a feature in cumulus binary. We are having mixed success with this. Its particularly worse since the relay chain block constantly lags beyond what is usually considered acceptable (>10 blocks) ever since v0.39. We struggle with > v0.39 for validators too. We are seriously considering downgrading back to v0.39 rather than following the convention of using the latest binary. 
What makes this hard is that the timeouts and block delays are infrequent and only a service restart seems to fix it. We will revert the external RPC to make sure that is not adding on to the timeout. We request our RPC to be added back. 

Meanwhile any advice on the matter of binary version is much appreciated.